Add files via upload
These are the webpages I have been working on. They still need work, but in the event of something catastrophic, we all will have access to them. They both need to redirect after completion and need CSS code to generate a more aesthetically pleasing webpage. In short, still needs work but on track with Sprint.