(resume) fix: convert_blob2text() works correctly
Previously I was attempting to pass the returned PDF binary data
directly to the parseFile() func, which actually expects a filename, not
raw binary input.
To accommodate this requirement, an intermediate step has been added to
place the returned binary PDF data into a tmp file on the server (
removed after use ), with that filename being passed to parseFile().

I also made the convert_blob2text() func private as it is only called
internally by Resume as of now.
Additionally I corrected an incorrect call to the function in the pull()
method.