(pdf_rednerer) testing out rendering pdfs
as of now, this is all placeholder/testing data to experiment with how
we would like render the pdf for printing.

TODO: determine how to extract the neccessary fields from the pdf
contents using the existing methods (and what would need to be built to
accomidate anything missing) to pass to the renderer.