(pdf_export) set up export to render before export
the export script can now be called with a resume id that the user would
like to render & download.

IMPORTANT NOTE:
while this is set up to work, as of now ( 2024-11-10 ), the render
function contains hardcoded elements to produce a test resume.

THIS IS SUBJECT TO CHANGE HOWEVER AS IS WILL PRODUCE ERRONEOUS BEHAVIOR!