add enum for resume styles
add enums to represent the style of a resume.
the idea here is to define an enum to assist with identifying a resume's
content's style to make parsing & rendering easier.
the enum has a public func `get_style_layout()` that returns an array of
headers for use during parsing.
there is the special case of a user defined resume style. right now this
returns an empty array and is therefore unimplemented in a production
way.
it might be best to simply not support custom formats.