[ end access refactor ] include session check & correct html->php calls
including the required session check in each important file.
currently these checks are not in the utility php code files ( the
classes etc ); that was causing errors. I will try to remedy that in the
future as those files should also be unaccessable to users via path.

changing the calls to *.hmtl to *.php in all of the necessary files.

adding redirect after resume upload ( this still needs work as its
functionality gets implemented/improved )