Uploading two files that are linked together
Front_end_create_resume.php is what the user will see when they click "create resume". They can input information for their resume and then submit it.
Once submitted, information is captured by createResume.php which hands the rendering of the resume. It outputs the resume.
At this time, it does not generate a downloadable copy for the user.
At this time it is still under development.