add functionality to edit resume
files altered:
- FrontEnd_createResume
- createResume

Giving users the ability to edit their resume after viewing it.
This is specifically helpful for AI content, but is generally a nice QoL
option.
The array is passed back using `$_SESSION`.
If `$_SESSION['info']` is set the array is entered into vars that are
displayed throughout the form.