refactor(Resume): pdf binaries->array for contents
Refactoring the resume objs to store the contents of a resume using an
array instead of as a binary & string of that bins content.
[ POTENTIALLY BREAKING ]
This effects a handful of functions but mainly is inconsequential on any
calls, however create & delete have had their signatures changed while
import has been removed.