refactor(resumes): manual saving -> auto saving
Moving from manual saving of resumes to auto saving.
This affects the following files:
- FrontEnd_createResume
- Resume
- createResume
- my_resumes

Generally speaking this refactor involves setting a new session item to
indicate if coming from 'my_resumes' and then handling the logic
accordingly.
This involves the pull/set of the resume w/ the DB, as well as
differentiating between a fully new & existing resumes during the
view/edit process.